SPAN LEAGUE APPLICATION INITIAL COMMIT
This commit adds the initial draft for The Span 
League Application is an application responsible 
to read a file that generally has an extension 
of .league, like the examples included in this 
repo in the league-examples-flies folder. 

*Note: The application accepts any text-based file 
with a different extension of .league, as long as 
contains a valid structure, the Span-League-Engine 
should be able to process the file.

The application is based on a maven multi-module project, 
which consists of three main projects which compose 
the application:

- SPAN-LEAGUE-ENGINE
- SPAN-LEAGUE-NON-UI
- SPAN-LEAGUE-UI

[LINK TO GIT-HUB]
https://github.com/Bigbossnake/SPAN-LEAGUE